Add new motion detection algorithm (DNE), refs #524
The "Do No Engage" algorithm is simply a rectangle specified by the user in
which the algorithm will ignore motion. The feature is implemented by allowing
the user to pass a Rectangle into the algorithm, and then the algorithm
only processes pixels that are not contained in the Rectangle.

This commit is a squash of pull request #524 by Kevin Kanzelmeyer (@kkanzelmeyer).

It was produced by cherry picking all changes into the master, squasing,
removing all changes that relates to pom.xml files (these were unnecessary)
and the ones related to GStramer 1.x driver and then finally reworking to address
comments from review.